stitching works
git-svn-id: file:///home/roberto/Git/svn@10549 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 982b406e4a073f572b3a07cbaa8bb435951eef89

+++ trunk/xmipp/libraries/graphicsj/Stitching/src/Interpolator.java
@@ -9,17 +9,17 @@
  */
 public class Interpolator {
 
-    public static float bilinear(float pixels[], int w, int h, double x, double y) {
+    public static double bilinear(float pixels[], int w, int h, double x, double y) {
         int x_ = (int) x;
         int y_ = (int) y;
 
         // @TODO Fix bounds
-        float v00 = (x_ >= 0 && x_ < w && y_ >= 0 && y_ < h) ? pixels[y_ * w + x_] : 0;
-        float v01 = (x_ + 1 >= 0 && x_ + 1 < w && y_ >= 0 && y_ < h) ? pixels[y_ * w + x_ + 1] : 0;
-        float v10 = (x_ >= 0 && x_ < w && y_ + 1 >= 0 && y_ + 1 < h) ? pixels[(y_ + 1) * w + x_] : 0;
-        float v11 = (x_ + 1 >= 0 && x_ + 1 < w && y_ + 1 >= 0 && y_ + 1 < h) ? pixels[(y_ + 1) * w + (x_ + 1)] : 0;
+        double v00 = (x_ >= 0 && x_ < w && y_ >= 0 && y_ < h) ? pixels[y_ * w + x_] : 0;
+        double v01 = (x_ + 1 >= 0 && x_ + 1 < w && y_ >= 0 && y_ < h) ? pixels[y_ * w + x_ + 1] : 0;
+        double v10 = (x_ >= 0 && x_ < w && y_ + 1 >= 0 && y_ + 1 < h) ? pixels[(y_ + 1) * w + x_] : 0;
+        double v11 = (x_ + 1 >= 0 && x_ + 1 < w && y_ + 1 >= 0 && y_ + 1 < h) ? pixels[(y_ + 1) * w + (x_ + 1)] : 0;
 
-        return bilinear(v00, v01, v10, v11, (float) x, (float) y);
+        return bilinear(v00, v01, v10, v11, (double) x, (double) y);
     }
 
     /**
@@ -31,9 +31,9 @@ public class Interpolator {
      * @param x,y point to interpolate
      * @return
      */
-    private static float bilinear(float v00, float v01, float v10, float v11, float x, float y) {
-        float v0 = linear(v00, v01, x);
-        float v1 = linear(v10, v11, x);
+    private static double bilinear(double v00, double v01, double v10, double v11, double x, double y) {
+        double v0 = linear(v00, v01, x);
+        double v1 = linear(v10, v11, x);
 
         return linear(v0, v1, y);
     }
@@ -45,10 +45,10 @@ public class Interpolator {
      * @param x value to interpolate
      * @return
      */
-    protected static float linear(float v0, float v1, float x) {
+    protected static double linear(double v0, double v1, double x) {
         int int_x = (int) x;
 
-        float d = x - int_x;
+        double d = x - int_x;
 
         return v1 * d - v0 * (d - 1);
     }
